requirements.txt
flask >= 0.11.1
Flask-OAuthlib >= 0.9.4
Flask-Session >= 0.3.0
mariadb-dyncol >= 1.2.0
mysqlclient >= 1.3.7
pytest >= 2.9.2
pytest-flask >= 0.10.0
pytest-watch >= 4.1.0
pyyaml
redis >= 2.10.5
requests >= 2.10.0
sqlalchemy >= 1.0.14

xlrd >= 1.0.0
openpyxl >= 2.4.1

python-dateutil >= 2.6.1

dumper >= 1.0.2

